refactor(datagrid): use native NSButton tristate checkbox for filter select-all
